+++ docs/CHANGELOG.md
@@ -12,6 +12,8 @@
 * Package `graph`:
   * `Gather` and `GatherSlices` now have and extra argument called `indicesAreSorted` that tells whether
     the start indices are guaranteed to be sorted, which allows some optimizations in some platforms.
+  * Exposed `BackendGather`, `BackendScatterMax`, `BackendScatterMin` and `BackendScatterSum` for test and debugging
+    purposes.
 * Moved code generation tools from `cmd` to `internal/cmd` directory.
 
 # v0.18.1: 2025/04/13 Many fixes, XLA update, Tensor clone.